+++ RawTools/WorkFlows/WorkflowParameters.cs
@@ -56,6 +56,7 @@ namespace RawTools.WorkFlows
             IncludeSubdirectories = (bool)TryGetElseDefault(Options,"SearchSubdirectories");
             RefineMassCharge = (bool)TryGetElseDefault(Options,"RefineMassCharge");
             ParseParams.Chromatogram = (string)TryGetElseDefault(Options, "Chromatogram");
+            ParseParams.WriteMgfLevels = (string)TryGetElseDefault(Options, "WriteMgfLevels");
             ParseParams.LabelingReagents = (string)TryGetElseDefault(Options, "LabelingReagent");
             ParseParams.Metrics = (bool)TryGetElseDefault(Options, "Metrics");
             ParseParams.Parse = (bool)TryGetElseDefault(Options, "Parse");
@@ -144,7 +145,7 @@ namespace RawTools.WorkFlows
     public class ParseWorkflowParameters
     {
         public bool Parse, Quant, UnlabeledQuant, WriteMgf, Metrics;
-        public string LabelingReagents, OutputDirectory, Chromatogram;
+        public string LabelingReagents, OutputDirectory, Chromatogram, WriteMgfLevels;
     }
 
     public class QcWorkflowParameters